ui/requirements.txt
pandas==2.2.0
streamlit==1.31.0
openai==1.12.0
watchdog==4.0.0
python-dotenv==1.0.1